// src/modules/inventory/reports/yarn-store/yarn-additional-booking-report-pdf/yarn-additional-booking-report-pdf-type.ts
export interface YarnAdditionalBookingReportType {
  ORDERPLACEMENTMONTH?: string | null;
  BOOKING_NO?: string | null;
  DISPLAY_NAME?: string | null;
  PONO?: string | null;
  FACTORY_NAME?: string | null;
  STYLENAME?: string | null;
  QUANTITY: number;
  APPROVAL_REMARKS?: string | null;
  PROBLEM_CATEGORY_NAME?: string | null;
  DEBIT_VOUCHER?: string | null;
  STATUS?: string | null;
  COMPANY_NAME?: string | null;
  COMPANY_ADDRESS?: string | null;
}